server/requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-environ==0.4.3
django-versatileimagefield==1.7.1
djangorestframework==3.6.3
olefile==0.44
Pillow==4.0.0
pytz==2017.2
six==1.10.0
